Fix: Modify electric_control.py
Modified python files - print functions are written in python2 style.
So I changed them to python3 style.

And add a test() function at the front of the main.
If input is 'y', the circuit connects and the light on.
Else if 'n', it makes the light off.
Else if 'b', it goes to next code and waiting for the beacon signal.